Move parseMarkdown to markdown
